requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai==0.8.5
Pillow
